backend/requirements.txt
# Essential Libraries
pandas==1.5.3
numpy==1.23.5
requests==2.28.1
scikit-learn==1.1.3
joblib==1.2.0

# Dash & Flask for Web Interface
dash==2.9.0
dash-core-components==2.0.0
dash-html-components==2.0.0
flask==2.2.2
flask-cors==3.0.10

# Database & Environment Handling
python-dotenv==1.0.0

# Machine Learning Models
xgboost==1.7.6
lightgbm==4.0.0
